requirements.txt
Django==2.2.11
djangorestframework==3.11.0
psycopg2==2.7.3.1
redis==3.4.1
django-redis==4.11.0
pytest==5.2.1
pytest-cov==2.8.1
pytest-factoryboy==2.0.3
factory-boy==2.12.0
pytest-django==3.5.1
pytest-docker-compose==3.1.2
django-extensions==2.2.8